src/requirements.txt
google_api_python_client==2.139.0
google_auth_oauthlib==1.2.1
jsonschema==4.23.0
Pillow==10.4.0
protobuf==5.27.3
reportlab==4.2.2
utils==1.0.2
